chore: move Single Stat stories and add custom csv option
